changed nextTile in Attacker from Dir enum to Tile
changed setMove in Tile to take in a boolean and the next tile

added the new parameter for setMove into Desertuprising so the map now
points to the next tile to go (like a linked list)

modified move() in attacker

added abstract getSpawnTile method to Map and put @Override over the
getSpawnTile in BaseDesertUprising

added Tile parameter to Attacker constructor so it can tell it's own
starting location easier

Started working on animating and moving the marines. Commented out what
I was working on there:
//startWaves in the GameController constructor
//a lot of stuff in the Ticker.update()

So it draws and moves (really fast because I didn't adjust it to get
along well with the ticker yet) and somehow one ends up on the wrong
path. Will look into that later